Support async JS embeds in React
